requirements.txt
Babel==2.7.0
captcha==0.3
certifi==2019.11.28
chardet==3.0.4
cmsplugin-filer==1.1.3
DateTimeRange==0.6.1
decorators==0.1.1
dj-database-url==0.5.0
Django==2.0
django-admin-ckeditor==0.0.1
django-appconf==1.0.3
django-bootstrap3==11.1.0
django-bootstrap3-datepicker==0.4.1
django-ckeditor==5.7.1
django-ckeditor-filebrowser-filer==0.3.0
django-ckeditor-staticfiles==4.3.3
django-classy-tags==0.9.0
django-cleanup==4.0.0
django-cms==3.7.0
django-datatables-view==1.19.0
django-filebrowser==3.12.1
django-filer==1.5.0
django-formtools==2.1
django-gamification==0.3.2
django-grappelli==2.13.3
django-jquery==3.1.0
django-js-asset==1.2.2
django-mailbox==4.8.1
django-mathfilters==0.4.0
django-meta==1.5.2
django-mptt==0.10.0
django-multiupload==0.5.2
django-phonenumber-field==3.0.1
django-polymorphic==2.0.3
django-recaptcha==2.0.5
django-sekizai==1.0.0
django-select2==7.1.1
django-summernote==0.8.11.6
django-tinymce==2.6.0
Django-tinymce-filebrowser==0.2.1
django-treebeard==4.3
djangocms-admin-style==1.4.0
djangocms-attributes-field==1.1.0
djangocms-column==1.9.0
djangocms-file==2.3.0
djangocms-googlemap==1.3.0
djangocms-icon==1.4.2
djangocms-link==2.5.0
djangocms-page-meta==0.9.0
djangocms-picture==2.3.0
djangocms-snippet==2.2.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.8.0
djangocms-video==2.2.0
djangorestframework==3.10.3
easy-thumbnails==2.6
fd==0.6
filelock==3.0.12
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
mbstrdecoder==0.8.1
meta==1.0.2
phonenumbers==8.10.23
pickleDB==0.9.2
Pillow==6.2.0
pinax-badges==2.0.0
pinax-points==1.0
psycopg2==2.7.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
reserve==0.3
sc-pylibs==0.1.4
simplejson==3.16.0
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
typepy==0.6.0
ui==0.1.4
Unidecode==1.0.23
urllib3==1.25.7
webencodings==0.5.1
whitenoise==4.1.4
